Split parser type declarations
